Allow running tests with PHPUnit 10 and 11
PHPUnit 11 emits complaints about the error handlers not being
removed, which is risky, so I had to add a few calls to
restore_error_handler().

This leaves us with PHPUnit deprecations about using @before and @after rather
than the corresponding attributes. I do not think those can be addressed
without a breaking change (adding the attributes is not enough, you have
to remove the annotations).